Refactors speech menu to make explorer independent from menu.
